refactor(dashboard): extract shared ResourceTable and migrate providers
